refactor [KVCacheManager, PageAllocator]: modularize locking mechanism by introducing a synchronized decorator and updating lock usage in PageAllocator for improved thread safety
